backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
anthropic==0.49.0
gitpython==3.1.41
pydantic==2.10.0
httpx==0.27.0
slowapi==0.1.9
